docs: update utility examples
